5% promotion added, bug fixes

+++ app/Http/Controllers/CourseController.php
@@ -62,17 +62,18 @@ class CourseController extends Controller
         if ($data["difficulty"]=="C1") $c->difficulty_id = 5;
         if ($data["difficulty"]=="C2") $c->difficulty_id = 6;
 
-        if ($data["form"]=="Stationary") $c->form_id = 1;
-        if ($data["form"]=="Remote") $c->form_id = 2;
-        if ($data["form"]=="Hybrid") $c->form_id = 3;
+        if ($data["form"]=="Remote") $c->form_id = 1;
+        if ($data["form"]=="Hybrid") $c->form_id = 2;
 
-        if ($data["img"]!=null) {
-            $max = Course::max("id")+1;
-            $name = $max.".png";
-            $c->img = $name;
+        if (isset($data["img"])) {
+            if ($data["img"]!=null) {
+                $max = Course::max("id")+1;
+                $name = $max.".png";
+                $c->img = $name;
 
-            if(Storage::exists('public/img/courses/'.$name)) Storage::delete('public/img/courses/'.$name);
-            Storage::putFileAs("public/img/courses", $data["img"], $name);
+                if(Storage::exists('public/img/courses/'.$name)) Storage::delete('public/img/courses/'.$name);
+                Storage::putFileAs("public/img/courses", $data["img"], $name);
+            }
         }
 
         $c->save();
@@ -97,14 +98,16 @@ class CourseController extends Controller
 
                 $already = false;
                 $payment = false;
+                $price = 0;
                 foreach ($enrollments as $e) {
                     if ($e->course_id == $id) {
                         $already = true;
+                        $price = $e->to_pay;
                         if ($e->payment_date != null) $payment = true;
                     }
                 }
 
-                return view("courses.show", ["c"=>$course, "already"=>$already, 'payment'=>$payment]);
+                return view("courses.show", ["c"=>$course, "already"=>$already, 'payment'=>$payment, "price"=>$price]);
             }
 
             if ($course->author_id == Auth::user()->id || Auth::user()->id == 1) {
@@ -173,9 +176,8 @@ class CourseController extends Controller
             if ($data["difficulty"]=="C1") $c->difficulty_id = 5;
             if ($data["difficulty"]=="C2") $c->difficulty_id = 6;
 
-            if ($data["form"]=="Stationary") $c->form_id = 1;
-            if ($data["form"]=="Remote") $c->form_id = 2;
-            if ($data["form"]=="Hybrid") $c->form_id = 3;
+            if ($data["form"]=="Remote") $c->form_id = 1;
+            if ($data["form"]=="Hybrid") $c->form_id = 2;
 
             $c->save();
